This was a fun lab to do. I did struggle a bit with the coding. I felt like I had to read and watch the links at least 5 tims in order to fully understand these Arrays/ArrayList, and all the Team and Search classes. I worked on it alone.
Q2:Answer: If the binarry wasnt sorted properly,Thew program would run w/ manyy errora.

Q3: Answer: So everything can be set up in a correct way for the rest of the Classes and methods to work.
Q4: Answer: The purpose of Final is so that the person doing the code doesnt, accidentaly change any thing that could ruin the program.